handle user authorization

+++ server/src/routes/memories.ts
@@ -3,41 +3,52 @@ import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async request => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async request => {
     const memories = await prisma.memory.findMany({
-      orderBy: {
-        createdAt: 'asc',
+      where: {
+        userId: request.user.sub
       },
+      orderBy: {
+        createdAt: 'asc'
+      }
     })
 
-    return memories.map((memory) => ({
+    return memories.map(memory => ({
       id: memory.id,
       coverUrl: memory.coverUrl,
-      excerpt: memory.content.substring(0, 120).concat('...'),
+      excerpt: memory.content.substring(0, 120).concat('...')
     }))
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, response) => {
     const paramsSchema = z.object({
-      id: z.string().uuid(),
+      id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(request.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
       where: {
-        id,
-      },
+        id
+      }
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return response.status(403).send()
+    }
 
     return memory
   })
 
-  app.post('/memories', async (request) => {
+  app.post('/memories', async request => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false),
+      isPublic: z.coerce.boolean().default(false)
     })
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
@@ -47,52 +58,72 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '',
-      },
+        userId: request.user.sub
+      }
     })
 
     return memory.id
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, response) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false),
+      isPublic: z.coerce.boolean().default(false)
     })
 
     const paramsSchema = z.object({
-      id: z.string().uuid(),
+      id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(request.params)
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
       where: {
-        id,
+        id
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return response.status(403).send()
+    }
+
+    memory = await prisma.memory.update({
+      where: {
+        id
       },
       data: {
         content,
         coverUrl,
-        isPublic,
-      },
+        isPublic
+      }
     })
 
     return memory.id
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, response) => {
     const paramsSchema = z.object({
-      id: z.string().uuid(),
+      id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(request.params)
 
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return response.status(403).send()
+    }
+
     await prisma.memory.delete({
       where: {
-        id,
-      },
+        id
+      }
     })
   })
 }
